-룸 스폰 위치 중앙으로 변경 -룸 크기 변경 이전 룸 크기: 15 x 30 x 15 (6750) 바뀐 룸 크기: 25 x 25 x 25  (15,625) -룸 Locker 제거 -room empty때 core false하게 버그수정 #-COoperative 미니게임 멤버는 토큰 지급 안되는 실수 고치기 #-FINISH굿즈를 re room finish 명령어로 합치기
#-re room load,empty명령어 쓸 때 단순히 BlockDatas만 로드되게 변경
-미니게임중 TP 사인 사용못하게 변경
-Bridge, Push, CENTER 배틀 미니게임 추가

+++ RelayEscapePlugin/src/com/wbm/plugin/util/general/TeleportTool.java
@@ -12,7 +12,7 @@ public class TeleportTool
 {
 	static Map<String, Location> locations;
 	
-	public static void tp(Entity src, int x, int y, int z) {
+	public static void tp(Entity src, double x, double y, double z) {
 		src.teleport(new Location(src.getWorld(), x, y, z));
 	}
